(cleanup) delete old profile component and rename to projectDashboard
