+++ client/app/app.js
@@ -22,7 +22,7 @@ import admin from './admin';
 import navbar from '../components/navbar/navbar.component';
 import footer from '../components/footer/footer.component';
 import main from './main/main.component';
-import profile from './profile/profile.component';
+import projectsDashboard from './projectsDashboard/projectsDashboard.component';
 import constants from './app.constants';
 import util from '../components/util/util.module';
 import socket from '../components/socket/socket.service';
@@ -30,7 +30,7 @@ import './app.scss';
 
 angular.module('startupbusApp', [ngCookies, ngResource, ngSanitize, 'btford.socket-io', uiRouter,
   uiBootstrap, _Auth, account, admin, 'validation.match', navbar, footer, main, constants,
-  socket, util, profile
+  socket, util, projectsDashboard
 ])
   .config(routeConfig)
   .run(function($rootScope, $location, Auth) {
