--- client/app/profile/profile.component.spec.js
@@ -1,17 +1,0 @@
-'use strict';
-
-describe('Component: ProfileComponent', function() {
-  // load the controller's module
-  beforeEach(module('startupBusApp.profile'));
-
-  var ProfileComponent;
-
-  // Initialize the controller and a mock scope
-  beforeEach(inject(function($componentController) {
-    ProfileComponent = $componentController('profile', {});
-  }));
-
-  it('should ...', function() {
-    expect(1).to.equal(1);
-  });
-});
