--- client/app/profile/profile.routes.js
@@ -1,10 +1,0 @@
-'use strict';
-
-export default function($stateProvider) {
-  'ngInject';
-  $stateProvider
-    .state('profile', {
-      url: '/profile',
-      template: '<profile></profile>'
-    });
-}
