--- client/app/profile/profile.scss
@@ -1,20 +1,0 @@
-  .profilePage{
-    background:purple;
-    height: 70rem;
-    padding: 3rem;
-    .profileInformation {
-      background: red;
-      height: 80%;
-    }
-    .photoAlbum {
-      background:green;
-      height: 80%
-    }
-    // .imagetable:{
-    //
-    // }
-    .image {
-      height:8rem;
-      margin:3px;
-    }
-  }
