+++ client/app/projectsDashboard/projectsDashboard.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<section message='message'  id="profilePage" class="profilePage row">
+<section id="profilePage" class="profilePage row">
   <section class="profileInformation col-md-4">
       <h1>THe profile header</h1>
   </section>
