+++ client/app/projectsDashboard/projectsDashboard.routes.js
@@ -1,0 +1,10 @@
+'use strict';
+
+export default function($stateProvider) {
+  'ngInject';
+  $stateProvider
+    .state('projectsDashboard', {
+      url: '/projectsDashboard',
+      template: '<projects-dashboard></projects-dashboard>'
+    });
+}
